Complete predicate move validation and pawn rules
En passant is not yet supported. The current method of determining
valid squares for pieces only evaluates properties of
pieces, not squares. This might need to be reworked when implementing
mechanics such as check, castling and en passant, since those require
knowledge of either the history of the board or the state of arbitrary
pieces.