Rework move validation, add Rook rules, instance Show Board
The previous two step move validation was overly complicated and was
therefore abolished. Instead of generating a predicate list to be
checked by some other function, the validMoves function is now
directly called with the relevant piece and the board. Each piece is
matched against the correct function. The Show implementation for
Board is not very robust, and should probably be cleaned up at some
point.