Change valid square finding process, impelment Bishop
Squares are now calculated with a two step process. The first pass
evaluates all squares that a piece targets. This is useful for
determining whether or not the king is allowed to move to a given
square or not. The second pass filters out the targeted squares that
are not valid moves. Pawn movement is an edge case, and has a modified
second pass.