Change check mate check, add display functions
Previous check mate check was faulty. The change necessitated
modifying the targeters function, which might lead to unexpected
behaviour if care is not taken. Targeters now also requires the caller
to provide the function which finds the target squares. This can
potentially lead to infinite recursion if the wrong function is
provided in the wrong context.