Implement serialization, and start securing comms
Communication relied heavily on trust between both hosts, with each
one having the ability to crash the others program in a lot of
cases. Work on securing the communication between them was started and
resulted in a lot of changes to the codebase, including the
introduction of existential quantification and friends.