Finish trustless verification of all incoming messages
All moves are checked before they are passed on to the board. If a
single move is invalid, the connection to the remote host is
immediately closed. If a host sends an invalid move, it can be
asserted that their client is dysfunctional or that they are acting in
bad faith. In both cases there is no real reason to resume.